requirements.txt
langchain>=0.1.14
langgraph>=0.0.24
google_generativeai
pydantic>=2.6
requests>=2.31
yfinance>=0.2.36
newsapi-python>=0.2.7
python-dotenv